+++ home/common/programs/default.nix
@@ -1,9 +1,9 @@
 {
+  ...
+}:
+{
   imports = [
-    ./browsers.nix
     ./common.nix
     ./git.nix
-    ./media.nix
-    ./xdg.nix
   ];
 }