+++ home/common/shell/core.nix
@@ -1,9 +1,10 @@
-{ username, ... }: {
+{ pkgs, username, ... }: {
   # Home Manager needs a bit of information about you and the
   # paths it should manage.
   home = {
     inherit username;
-    homeDirectory = "/home/${username}";
+    homeDirectory =
+      if pkgs.stdenv.isDarwin then "/Users/${username}" else "/home/${username}";
 
     # This value determines the Home Manager release that your
     # configuration is compatible with. This helps avoid breakage