+++ home/linux/programs/default.nix
@@ -1,0 +1,8 @@
+{
+  imports = [
+    ./browsers.nix
+    ./common.nix
+    ./media.nix
+    ./xdg.nix
+  ];
+}